Split gcp client module into gce and gcrm
In order to follow convention of 1 client per module.
