+++ gordon_janitor_gcp/clients/__init__.py
@@ -16,14 +16,16 @@
 
 # Mainly for easier documentation reading
 from gordon_janitor_gcp.clients.auth import *  # noqa: F403
-from gordon_janitor_gcp.clients.gcp import *  # noqa: F403
+from gordon_janitor_gcp.clients.gce import *  # noqa: F403
+from gordon_janitor_gcp.clients.gcrm import *  # noqa: F403
 from gordon_janitor_gcp.clients.gdns import *  # noqa: F403
 from gordon_janitor_gcp.clients.http import *  # noqa: F403
 
 
 __all__ = (
     auth.__all__ +  # noqa: F405
-    gcp.__all__ +  # noqa: F405
+    gce.__all__ +  # noqa: F405
+    gcrm.__all__ +  # noqa: F405
     gdns.__all__ +  # noqa: F405
     http.__all__  # noqa: F405
 )
